packages/bsb-core/pyproject.toml
[build-system]
requires = [ "flit_core >=3.2,<4" ]
build-backend = "flit_core.buildapi"

[project]
name = "bsb-core"
version = "6.0.7"
readme = "README.md"
requires-python = ">=3.10,<4"
dynamic = [ "description" ]
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
dependencies = [
  "numpy~=1.19",
  "scipy~=1.5",
  "scikit-learn~=1.0",
  "errr~=1.2",
  "rtree~=1.0",
  "psutil~=5.8",
  "voxcell~=3.1",
  "morphio~=3.3",
  "toml~=0.10.0",
  "requests~=2.30",
  "urllib3~=2.0",
  "appdirs~=1.4",
  "neo[nixio]~=0.13",
  "tqdm~=4.50",
  "dashing-next~=0.1",
  "shortuuid~=1.0",
  "quantities~=0.15.0",
  "exceptiongroup>=1.2.0",
  "opentelemetry-api~=1.0"
]

  [[project.authors]]
  name = "Robin De Schepper"
  email = "robin@alexandria.sc"

  [[project.authors]]
  name = "Dimitri Rodarie"
  email = "dimitri.rodarie@unipv.it"

  [project.license]
  file = "LICENSE"

  [project.optional-dependencies]
  parallel = [ "mpi4py~=3.0", "mpipool>=2.2.1,<3", "mpilock~=1.1" ]

[project.entry-points."bsb.storage.engines"]
fs = "bsb.storage.fs"

[project.entry-points."bsb.commands"]
commands = "bsb.cli.commands._commands"
projects = "bsb.cli.commands._projects"

[project.entry-points."bsb.options"]
verbosity = "bsb._options:verbosity"
sudo = "bsb._options:sudo"
version = "bsb._options:version"
config = "bsb._options:config"
sim_console_progress= "bsb._options:sim_console_progress"
profiling = "bsb._options:profiling"
debug_pool = "bsb._options:debug_pool"
quiet = "bsb._options:quiet"

  [project.scripts]
  bsb = "bsb.cli:handle_cli"

[dependency-groups]
test = [
  "bsb-core[parallel]",
  "bsb-arbor~=6.0; sys_platform != 'win32'",
  "bsb-hdf5~=6.0",
  "bsb-json~=6.0",
  "bsb-test~=6.0",
  "coverage>=7.3"
]
docs = [
  "furo~=2024.4",
  "sphinxemoji~=0.3",
  "sphinx-copybutton~=0.5",
  "bsb-json~=6.0",
  "bsb-yaml~=6.0",
  "sphinxext-bsb~=6.0"
]
dev = [
  "pre-commit~=3.5",
  "ruff>=0.8.2",
  "snakeviz~=2.1",
  "opentelemetry-distro>=0.59b0",
  "opentelemetry-exporter-otlp~=1.38",
]

[tool.uv]
default-groups = [ "dev", "docs", "test" ]

[tool.uv.sources.bsb-test]
path = "../bsb-test"
editable = true

[tool.uv.sources.bsb-hdf5]
path = "../bsb-hdf5"
editable = true

[tool.uv.sources.bsb-arbor]
path = "../bsb-arbor"
editable = true

[tool.uv.sources.bsb-json]
path = "../bsb-json"
editable = true

[tool.uv.sources.bsb-yaml]
path = "../bsb-yaml"
editable = true

[tool.uv.sources.sphinxext-bsb]
path = "../../libs/sphinxext-bsb"
editable = true

[tool.flit.module]
name = "bsb"

[tool.coverage.run]
branch = true
source = [ "bsb" ]

[tool.coverage.report]
exclude_lines = [
  "if typing.TYPE_CHECKING:",
  "pragma: nocover"
]
show_missing = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 90
indent-width = 4

  [tool.ruff.format]
  quote-style = "double"
  indent-style = "space"
  skip-magic-trailing-comma = false
  line-ending = "auto"
  docstring-code-format = true
  docstring-code-line-length = 90

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ "F822", "B019" ]
  fixable = [ "ALL" ]
  unfixable = [ ]

    [tool.ruff.lint.per-file-ignores]
    "bsb/__init__.py" = [ "E501" ]
